backend/pyproject.toml
[project]
name = "backend"
version = "0.0.0"
description = "Backend"
readme = "README.md"
requires-python = ">=3.13.0,<3.14.0"
dependencies = [
    "aiohttp>=3.13.1",
    "argon2-cffi>=23.1.0",
    "dj-database-url>=2.3.0",
    "django>=5.2.7",
    "django-allauth[socialaccount]==65.7.0",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.3",
    "gunicorn>=23.0.0",
    "pymupdf>=1.24.0",
    "psycopg[binary]>=3.2.0",
    "requests>=2.32.0",
    "uncouple>=1.0.0",
    "whitenoise>=6.11.0",
    "langgraph>=1.0.3",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "langchain-openai>=1.0.3",
    "pgvector>=0.3.8",
]

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "pytest>=8.4.2",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.0",
    "werkzeug>=3.1.3",
]
